Persistência no banco

+++ back/src/servicos/distanciasGeograficas/ConfiguracoesDistanciasGeograficas.java
@@ -1,6 +1,6 @@
 package servicos.distanciasGeograficas;
 
 public class ConfiguracoesDistanciasGeograficas {
-	public static final int RAIO_DA_TERRA = 6371;
+	public static final double RAIO_DA_TERRA = 6371;
 
 }
